+++ R/geom_tiplab.R
@@ -46,10 +46,12 @@ geom_tiplab_rectangular <- function(mapping=NULL, hjust = 0,  align = FALSE, lin
 
     x <- y <- label <- isTip <- node <- NULL
     if (align == TRUE) {
-        self_mapping <- aes(x = max(x, na.rm=TRUE) + diff(range(x, na.rm=TRUE))/200, y = y, label = label, node = node, subset = isTip)
+        self_mapping <- aes(x = max(x, na.rm=TRUE) + diff(range(x, na.rm=TRUE))/200, y = y,
+                            label = label, node = node, subset = isTip)
     }
     else {
-        self_mapping <- aes(x = x + diff(range(x, na.rm=TRUE))/200, y= y, label = label, node = node, subset = isTip)
+        self_mapping <- aes(x = x + diff(range(x, na.rm=TRUE))/200, y= y,
+                            label = label, node = node, subset = isTip)
     }
 
     if (is.null(mapping)) {
@@ -78,9 +80,13 @@ geom_tiplab_rectangular <- function(mapping=NULL, hjust = 0,  align = FALSE, lin
                           linetype = linetype, nudge_x = offset,
                           size = linesize, stat = StatTreeData, ...)
        ,
-        label_geom(mapping=text_mapping,
-                   hjust = hjust, nudge_x = offset, stat = StatTreeData, fontface = fontface, ...)
-
+        if (geom %in% c("image", "phylopic")) {
+            label_geom(mapping=text_mapping,
+                       hjust = hjust, nudge_x = offset, stat = StatTreeData, ...)            
+        } else {
+            label_geom(mapping=text_mapping,
+                       hjust = hjust, nudge_x = offset, stat = StatTreeData, fontface = fontface, ...)
+        }
     )
 }